feat: implement new Menu component with enhanced features and structure
- Introduced a new Menu component that replaces the previous MenuV2, featuring improved item rendering and search functionality.
- Added MenuGroupLabel and MenuGroupSeperator for better organization of menu items.
- Implemented dynamic filtering of menu items based on search input.
- Enhanced MenuItem and SubMenu components for better usability and styling.
- Updated main export file to include the new Menu component and its related types.
- Refactored MenuDemo to showcase the new Menu component and its features.
