+++ lib/components/Menu/index.ts
@@ -1,0 +1,2 @@
+export { default as Menu } from "./Menu";
+export * from "./types";
