+++ lib/components/Select/index.ts
@@ -1,0 +1,2 @@
+export { default as Select } from "./Select";
+export * from "./types";
