+++ lib/main.ts
@@ -22,3 +22,5 @@ export * from "./components/Snackbar";
 export * from "./components/DateRangePicker"
 export * from "./components/StatCard";
 export * from "./components/Inputs";
+export * from "./components/Select";
+export * from "./components/Menu";
