+++ src/demos/Menu/MenuDemo.tsx
@@ -1,78 +1,10 @@
-import {
-  ChevronRightIcon,
-  CopyIcon,
-  FileIcon,
-  PlusIcon,
-  SettingsIcon,
-  User,
-} from "lucide-react";
+import { User } from "lucide-react";
 import { Button } from "../../../lib/components/Button";
-import {
-  MenuGroupType,
-  MenuItemActionType,
-  MenuItemVariant,
-} from "../../../lib/components/Menu/Menu";
 import Block from "../../../lib/components/Primitives/Block/Block";
-import MenuV2 from "../../../lib/components/MenuV2/MenuV2";
+import Menu from "../../../lib/components/Menu/Menu";
 import Select from "../../../lib/components/Select/Select";
 import { useState } from "react";
-import { MenuAlignment } from "../../../lib/components/MenuV2/types";
-
-const dummyItems: MenuGroupType[] = [
-  {
-    label: "Options",
-    showSeparator: true,
-    items: [
-      {
-        variant: MenuItemVariant.DEFAULT,
-        text: "Release Now",
-        slot4: <kbd>⌘</kbd>,
-        onClick: () => alert("Edit clicked"),
-      },
-      {
-        variant: MenuItemVariant.ACTION,
-        slot1: <FileIcon size={13} />,
-        text: "Save As CSV",
-        onClick: () => alert("Delete clicked"),
-      },
-    ],
-  },
-  {
-    items: [
-      {
-        variant: MenuItemVariant.ACTION,
-        actionType: MenuItemActionType.PRIMARY,
-        slot1: <PlusIcon size={14} color={"blue"} />,
-        text: "Create New",
-        onClick: () => alert("Select All clicked"),
-      },
-      {
-        text: "Copy CSV",
-        slot1: <CopyIcon size={13} />,
-        onClick: () => alert("This won't fire"),
-      },
-      {
-        text: "Merchant Settings",
-        slot1: <SettingsIcon size={13} />,
-        slot2: <ChevronRightIcon size={13} />,
-        subMenu: {
-          items: [
-            {
-              variant: MenuItemVariant.DEFAULT,
-              text: "Edit",
-              onClick: () => alert("Edit clicked"),
-            },
-            {
-              variant: MenuItemVariant.ACTION,
-              text: "Delete",
-              onClick: () => alert("Delete clicked"),
-            },
-          ],
-        },
-      },
-    ],
-  },
-];
+import { MenuAlignment } from "../../../lib/components/Menu/types";
 
 const MenuDemo = () => {
   // Add state for Select demo
@@ -89,7 +21,7 @@ const MenuDemo = () => {
         gap={100}
         style={{ marginTop: "20px" }}
       >
-        <MenuV2
+        <Menu
           alignment={MenuAlignment.END}
           enableSearch={false}
           trigger={<Button text="Open MenuV2" onClick={() => {}} />}
